+++ tests/testthat/test-metadata.R
@@ -37,8 +37,7 @@ test_that("inputs are the expected names and classes", {
       spreadFitAdditionalColNames = "character",
       standAgeMaps                = "list",
       studyArea                   = "SpatVector",
-      studyArea_biomassParam      = "SpatVector",
-      studyAreaReporting          = "sf")
+      studyArea_biomassParam      = "SpatVector")
   )
 })
 
@@ -54,7 +53,6 @@ test_that("outputs are the expected names and classes", {
       fireSense_escapeCovariates             = "data.table",
       fireSense_escapeFormula                = "character",
       fireSense_ignitionCovariates           = "data.table",
-      fireSense_ignitionFormula              = "character",
       fireSense_nonAnnualSpreadFitCovariates = "list",
       fireSense_spreadFormula                = "character",
       flammableRTM                           = "SpatRaster",
@@ -88,13 +86,13 @@ test_that("parameters are the expected names", {
   md <- SpaDES.core::moduleMetadata(module = moduleName, path = modulePath)
   expect_identical(
     sort(md$parameters$paramName),
-    sort(c(".plotInterval", ".saveInitialTime", ".saveInterval", ".studyAreaName",
+    sort(c(".studyAreaName",
            ".useCache", ".useCacheArgs", "areaMultiplier", "bufferForFireRaster", "cutoffForYoungAge",
            "dataYears", "estimateFuelClasses", "fireYears", "flammabilityThreshold",
-           "forestedLCC", "fuelClassCol", "igAggFactor", "igFocalFactor",
-           "minBufferSize", "modelAlgorithm", "nonflammableLCC",
+           "forestedLCC", "fuelClassCol", "igAggFactor",
+           "minBufferSize", "nonflammableLCC",
            "nonForestCanBeYoungAge", "sppEquivCol", "spreadFitFilename",
-           "spreadFitGoogleDriveFolder", "targetFuelClasses", "useCentroids",
+           "spreadFitGoogleDriveFolder", "targetFuelClasses",
            "useRasterizedFireForSpread", "whichModulesToPrepare"))
   )
 })